Fixed result check from send() call; in C the return would be negative, but in Perl it's undef in case of error.
Realized that the reason recv() was seeing no error when the syscall was
interrupted was because the SIGALRM handler's send() call stomps on the
contents of $!. Back it up and restore it, so that the error code check
can actually be useful. Also removed the check for 'if ($! == 0) ...', since
that's now irrelevant.

Added a '-b' switch to enable broadcast mode on the socket - 0.255 broadcast
pings can occasionally be interesting...