Remodeled how incoming request queue is stored slightly; rather than having a queue of transaction IDs, the actual request control blocks are queued.
Improved the debugging print statements to use caller() to get the function
name (with package) instead of hard coding.

Added a semaphore for the incoming request queue, and made changes to
GetTransaction() to make it possible to filter the incoming request
queue using a subroutine ref to make the choice, and to block until a
transaction (of the desired type, if needed) arrives.

More FIXMEs.