Commented out most of the debugging output, after some testing.
Realized I'd forgotten to supply the target address in the transaction resend
code; made the necessary additions to store the target sockaddr_at and
use it in the send() call.

Added code to signal any pending waiting transactions of failure when the
thread goes away.

Fixed up sharing in the RespondTransaction() callback for RspCB block setup.