Added fold marks.
Changed the comment at the start of the module to more accurately represent
the current state of the code.

Dropped inclusion of IO::Poll and threads from Net::Atalk::ASP, as it
doesn't use either of the two directly.

Removed a few error code definitions that are best referenced from
Net::AFP::Result; I want to move the rest of the error codes there as well.