Improved option handling in SendTransaction().
Improved internal naming consistency for several variables and hash
elements, to make things a little more interchangeable in the future,
if needed.

Changed the use of gettimeofday() so that we don't have to manually do
the math to get microseconds in their appropriate place, and changed the
hash element storage to store it as a single value instead of separate
second and microsecond values.

Altered code paths in the thread in Net::Atalk::ATP to only update the
sequence mask in the raw packet data for an outgoing transaction when we
actually have to resend the packet data.

Added a call to up() the incoming request queue semaphore on thread exit
in Net::Atalk::ATP.

Removed old commented invocations of SendTransaction() in Net::Atalk::ASP.

Changed the sequence wrap values in Net::Atalk::ASP from 65536 to
2 ** 16; they're technically the same value, but I think it's better to
represent it that way anyway.