Set the TCP_NODELAY flag on the TCP socket in Net::DSI; this substantially increases performance and responsiveness when using the AFP file share.
In Net::AFP::Fuse, pass the read *and* write bits when a file open is
requested to be write-only. Unfortunately, it seems that AirDisk breaks
when you pass only the write bit (you can't write to the open handle,
don't ask me why). It appears Apple's client does this too.