Changed Net::Atalk::ATP to add an interface to allow registering a callback that is just called at some given time interval. Also added a check to SendTransaction() to check the 'running' flag, and if the thread isn't running, to return kASPSessClosed.
In Net::Atalk::ASP, changed the _TickleFilter subroutine to keep a
timestamp for the last time a Tickle packet was received, and found
that the callback wasn't actually getting called. Turns out it's sent
from the session port, not the server port. Fixed that up, and added
a periodic function to check the last tickle timestamp and tell the
thread to exit if more than SP_TIMEOUT seconds have elapsed. Also
altered all SP* operations to return the value in $sem (from
SendTransaction()) immediately instead of trying to blindly call
the down() method on a semaphore object which isn't there.