Altered Net::DSI inner loop to do some processing of the DSI message packet before reading the payload, and to read directly into the scalar buffer passed instead of reading into a local buffer and copying out (except for OP_DSI_ATTENTION).
Fixed Net::AFP::FPLoginCont() to deal better with the buffer passed in
remaining 'undef' after calling SendAFPMessage().

Changed Net::AFP::TCP to pass an attention quantum value in the
call to DSIOpenSession().

Altered Net::AFP::Fuse to conditionally pull in the AppleTalk support
modules. Also changed inclusions to only pull in Fcntl once, moved the
decision of functions for reading, writing and enumerating files into
new() instead of conditionals later on, and changed the read() method
to return undef instead of -ESPIPE if the read returns no data and an
EOF error.

Changed afpsh.pl to get the character encoding for the current locale,
conditionalize ACL checking on the presence of the kSupportsACLs flag,
and moved the decision on functions for reading, writing and enumeration
to early on in the script startup rather than deciding in
conditionals later in the code. Also started running read lines through
decode() with the character encoding we get earlier in the program. Also
added transliteration from / characters in filenames to : chars and
vice versa in appropriate locations.