Added Debian packaging infrastructure.
Moved the Net/ subdirectory under lib/, to be consistent with the typical
Perl module structure.

Added a Makefile.PL for easier installation, and a Todo file containing
ideas on what needs doing.

Retooling of several methods in Net::AFP, to use key-value pair based
arguments instead of long lists of sequential arguments. I've updated the
inline docs for some of these, but not all.

Switched to using spaces instead of tabs in a lot of places.

Various bug fixes.

Changed afpmount.pl to create the Net::AFP::Fuse object in an eval block,
since a few things (like not being able to parse the AFP URL) will make it
throw an exception. Also check the arguments a bit better before trying to
use them. Also also, accept a -o switch at the command line, but just
ignore the option string if one is passed. This makes it easier to then
mount an AFP URL from fstab.

Moved the PDF documents out of the trunk.

Added support to send UTF-8 names as decomposed Unicode, and convert them
to composed Unicode when we receive them. Nautilus doesn't seem to mind
decomposed UTF8, but the command line is a while other story.

Added real usage information for afpmount.pl.