Fixed the DHX2 and DHCAST128 UAMs to send the NewPassword field before the OldPassword field instead of after in ChangePassword(), seems I somehow misread that.
Added a ChangePassword() wrapper in Net::AFP::UAMs, so we can actually do
password changing with an AFP server.

Changed the UAM search method to find the include path that we're pulling
Net::AFP::UAMs from, and to use the UAM packages that appear in the 'UAMs'
directory in the same base path.