Reorganized some imports.
Eliminated the 'ofcache' for tracking open files. Now returning an actual
file descriptor (the "open fork reference number" returned from
FPOpenFork()) c/o an improvement to the Fuse module.

Eliminated write coalescing; use the big_writes mount option instead.

Added UID/GID mapping, to translate the UID/GID associated with the
user we connected to the AFP server under to our (or a designated)
UID/GID on this end. Also ingesting options via new() to optionally get
user-specified UID/GID for mapping purposes.

Added a callout to chmod() in the mknod() implementation, to set the
permissions on the file to those provided.

Eliminated FPAccess() checks within the read() and write() implementations,
instead doing the necessary checking in the open() implementation.

Altered listxattr() and getxattr() implementations to do two of
FPListExtAttrs() and FPGetExtAttr(), respectively; one to request the
size of the data to be returned, and one to get the actual data. We were
previously sending a fixed size of 131072, which had worked to date, but
I couldn't guarantee it would always work.

Added implementations of opendir(), readdir(), releasedir(), fsyncdir(),
access(), create(), ftruncate(), fgetattr(), lock(), utimens(), and bmap()
based on my own improvements to the Perl Fuse module.