Only print out a line when lock() is called when debugging is enabled.
Just call getattr() when fgetattr() is called if possible, for now. Seems
Jaffer is tarded.

Remove the 'print' call that was outputing the return code from the call
to FPMoveAndRename.

In the unlink() implementation, also ignore the FPAccess() return if
the return code is kFPParamErr.

Correct a comment in the open() implementation.