Don't import Socket6 into local namespace.
Perl 5.12.x (or at least its Socket module) adds some symbols that conflict
with Socket6 (not enough that we could just not use Socket6, of course).
To work around that, don't do Socket6->import(); instead, just prefix
the symbols that need to come out of the Socket6 namespace with the
namespace prefix, and prefix with & and postfix with () to make sure the
Perl interpreter calls them as subs, and doesn't pass the caller's arglist.